# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: huMANity API - Explore history and culture of nations
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /humanity-timeline:
    get:
      operationId: getHumanityTimeline
      tags: [timeline]
      summary: Cross-civilization humanity timeline
      description: Returns events from different civilizations grouped by historical era
      parameters:
        - name: era
          in: query
          required: false
          schema:
            type: string
            enum: [ancient, classical, medieval, modern, contemporary]
      responses:
        "200":
          description: Humanity timeline by era
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumanityTimelineEra"
  /humanity-timeline/search:
    get:
      operationId: searchTimeline
      tags: [timeline]
      summary: Search timeline events across all nations
      description: Diacritic-insensitive search across nation names, event titles, and descriptions, with optional category/era filters.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search term (matches nation name, title, description)
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by event category
        - name: era
          in: query
          required: false
          schema:
            type: string
            enum: [ancient, classical, medieval, modern, contemporary]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineSearchResult"
  /humanity-timeline/compare:
    get:
      operationId: compareTimelines
      tags: [timeline]
      summary: Compare the timelines of multiple nations side by side
      parameters:
        - name: countries
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated country codes (2-4)
      responses:
        "200":
          description: Per-nation timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NationTimeline"
...
components:
  schemas:
...
    TimelineEvent:
      type: object
      required: [id, countryCode, year, title, description]
      properties:
        id:
          type: integer
        countryCode:
          type: string
        year:
          type: integer
        title:
          type: string
        description:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
...
    HumanityTimelineEra:
      type: object
      required: [era, label, yearRange, events]
      properties:
        era:
          type: string
        label:
          type: string
        yearRange:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
    TimelineSearchResult:
      type: object
      required: [id, countryCode, countryName, year, title, description]
      properties:
        id:
          type: integer
        countryCode:
          type: string
        countryName:
          type: string
        countryFlagUrl:
          type: ["string", "null"]
        year:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: ["string", "null"]
        era:
          type: string
    NationTimeline:
      type: object
      required: [code, name, events]
      properties:
        code:
          type: string
        name:
          type: string
        flagUrl:
          type: ["string", "null"]
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
...